+++ tests/PluginTest.php
@@ -1,0 +1,29 @@
+<?php declare(strict_types=1);
+/**
+ *
+ * class FfuenfNoWishlist
+ *
+ * @category   Shopware
+ * @package    Shopware\Plugins\FfuenfNoWishlist
+ * @author     Achim Rosenhagen / ffuenf - Pra & Rosenhagen GbR
+ * @copyright  Copyright (c) 2018, Achim Rosenhagen / ffuenf - Pra & Rosenhagen GbR (https://www.ffuenf.de)
+ *
+ */
+
+namespace FfuenfNoWishlist\Tests;
+
+use FfuenfNoWishlist\FfuenfNoWishlist as Plugin;
+use Shopware\Components\Test\Plugin\TestCase;
+
+class PluginTest extends TestCase
+{
+    protected static $ensureLoadedPlugins = [
+        'FfuenfNoWishlist' => [],
+    ];
+    public function testCanCreateInstance()
+    {
+        /** @var Plugin $plugin */
+        $plugin = Shopware()->Container()->get('kernel')->getPlugins()['FfuenfNoWishlist'];
+        $this->assertInstanceOf(Plugin::class, $plugin);
+    }
+}