+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of FfuenfNoWishlist.
 
+## 1.0.1 (December 7, 2018)
+
+* move ConfigAbstract to AbstractService
+* integrate test-suite
 
 ## 1.0.0 (December 4, 2018)